app/Sender/serializer.py
from rest_framework import serializers
from .models import *
from django.contrib.auth.models import User

class SenderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Sender
        fields = '__all__'

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        exclude = ('password', )
    