app/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
pika==1.3.1
requests==2.25.1
django-cors-headers==3.13.0
tzdata==2022.7
django-environ==0.9.0
# mysqlclient